Do some more refactoring to get the bags API tests passing
