+++ storage/bags_api/src/main/scala/uk/ac/wellcome/platform/storage/bags/api/BagsApi.scala
@@ -8,6 +8,7 @@ import akka.stream.ActorMaterializer
 import grizzled.slf4j.Logging
 import uk.ac.wellcome.Runnable
 import uk.ac.wellcome.platform.archive.common.config.models.HTTPServerConfig
+import uk.ac.wellcome.platform.archive.common.http.{HttpMetrics, WellcomeHttpApp}
 import uk.ac.wellcome.platform.archive.registrar.common.models.StorageManifest
 import uk.ac.wellcome.storage.ObjectStore
 import uk.ac.wellcome.storage.vhs.{EmptyMetadata, VersionedHybridStore}
@@ -18,6 +19,7 @@ class BagsApi(
   vhs: VersionedHybridStore[StorageManifest,
                             EmptyMetadata,
                             ObjectStore[StorageManifest]],
+  httpMetrics: HttpMetrics,
   httpServerConfig: HTTPServerConfig,
   contextURL: URL
 )(implicit val actorSystem: ActorSystem,
@@ -25,20 +27,19 @@ class BagsApi(
   ec: ExecutionContext)
     extends Logging
     with Runnable {
+
   val router = new Router(
     vhs = vhs,
     contextURL = contextURL
   )
 
-  val bindingFuture: Future[Http.ServerBinding] = Http()
-    .bindAndHandle(router.routes, httpServerConfig.host, httpServerConfig.port)
+  val app = new WellcomeHttpApp(
+    routes = router.routes,
+    httpMetrics = httpMetrics,
+    httpServerConfig = httpServerConfig,
+    contextURL = contextURL
+  )
 
   def run(): Future[Http.HttpTerminated] =
-    bindingFuture
-      .map(b => {
-        info(s"Listening on ${httpServerConfig.host}:${httpServerConfig.port}")
-
-        b
-      })
-      .flatMap(_.whenTerminated)
+    app.run()
 }
